Create product table and model
The price column is implemented as integer, as that seems the most
straightforward solution in this case. Decimals is sometimes encouraged
or even the use of the Money-Rails gem. It's important to not have it as
float.

Also, we want to make sure the price is not negative and that the code
is unique. The enum adds additional validations to ensure that at this
stage we only support 3 types of products.