Create basket model and relationship to products
The relationship to products is set up as a has_many through:
relationship. It gives us more flexibility in the future, allows us to
use the join model in isolation for callbacks, and furthermore models that
a product can belong to many baskets and a basket can have many
products.