POST basket endpoint
This adds a basket controller and a create action, which means a basket
can be created with a POST request.

The controller adds a list of products from the request to the basket.
The logic for that lives on the basket model. The product

The tests cover different scnenarios of what products being added and
the right price being returned.