Create DiscountService
The logic for calculating how discounts are applied is extracted into a
this service. It takes a basket object and then writes the discounted
price in the Basket#checkout_price column.

Tshirts get a straighforward 30% discount when buying 3 or more.
Getting the discount for mugs is a bit more involved.

Spelled out the discount for mugs works as follows:

2% discount for 10 to 19 items
4% discount for 20 to 29 items
6% discount for 30 to 39 items
8% discount for 40 to 49 items
10% discount for 50 to 59 items
12% discount for 60 to 69 items
14% discount for 70 to 79 items
16% discount for 80 to 89 items
18% discount for 90 to 99 items
20% discount for 100 to 109 items
22% discount for 110 to 119 items
24% discount for 120 to 129 items
26% discount for 130 to 139 items
28% discount for 140 to 149 items
30% discount for 150 or more items

We could add these discounts to a hash or even a table and then lookup
the number of mugs and find the right discount in each case. However,
they all follow a pattern, namely that the percentage is always 5 times
the mininum amount of mugs required to apply for this discount.
E.g. 150 is five times 30, 14 is 5 times 70 and so forth. Therefore we
were able to calculate the right discount on the fly in the
`calculate_mug_discount` method.